moved some code from model/navbar.js to components/navbar.js
