+++ src/main.js
@@ -1,8 +1,5 @@
 // import { Test } from './components/test.js'
 import { NavBar } from './components/NavBar/NavBar.js'
-import { NavBarModel } from './model/NavBar/main.js'
 /* eslint-disable */
 new NavBar(document.querySelector('.nav-bar'))
-new NavBarModel(document.querySelector('.nav-bar'))
-// new Test(document.querySelector('.wrapper'))
 /* eslint-enable */
