+++ webpack.dev.js
@@ -1,7 +1,6 @@
 const ExtractTextPlugin = require('extract-text-webpack-plugin')
 const merge = require('webpack-merge')
 const common = require('./webpack.common.js')
-console.log('holi estoy en dev')
 
 module.exports = merge(common, {
   devtool: 'inline-source-map',
